refactor(codegen): extrai passes/hoist_fn.rs + passes/destructuring.rs
- passes/hoist_fn.rs (478 LOC) — hoista Expr::Fn / Expr::Arrow para
  Item::Function sintetico + descend recursivo + visit_stmt.
- passes/destructuring.rs (581 LOC) — expand_destructuring com helpers
  make_const_decl/make_index_access/make_member_access/with_default,
  push_object_rest_decls e push_array_rest_decl.

func.rs: 6255 -> 5194 linhas. Suite verde (11/11 crates).
