+++ crates/rts-codegen/src/codegen/lower/func.rs
@@ -30,6 +30,8 @@ use super::passes::args::default_args::expand_default_args;
 use super::passes::args::rest_args::expand_rest_args;
 use super::passes::args::spread_args::expand_spread_args;
 use super::passes::async_expand::{expand_async_functions, expand_await_exprs};
+use super::passes::destructuring::expand_destructuring;
+use super::passes::hoist_fn::hoist_fn_expressions;
 use super::passes::object_methods::desugar_object_methods;
 use super::passes::static_fields::expand_static_fields;
 use super::ctx::{ClassMeta, FnCtx, GlobalVar, UserFnAbi, ValTy};
@@ -1730,1069 +1732,6 @@ struct LiftAcc {
     needs_c_callconv: HashSet<String>,
 }
 
-/// Top-level \`const x = function() {}\` ja foi convertido pelo parser
-/// (try_lower_fn_expr_decl); este pass cobre o que sobrou. Nao tenta
-/// resolver capturas de escopo — se o body referenciar variaveis
-/// fora dele, falha em codegen como ident undefined.
-fn hoist_fn_expressions(program: &mut Program) {
-    use swc_ecma_ast::{ArrowExpr, BlockStmtOrExpr};
-
-    let mut counter: u32 = 0;
-    let mut new_fns: Vec<Item> = Vec::new();
-
-    fn fresh_ident(name: &str) -> Expr {
-        Expr::Ident(swc_ecma_ast::Ident {
-            span: Default::default(),
-            ctxt: Default::default(),
-            sym: name.into(),
-            optional: false,
-        })
-    }
-
-    fn pat_to_param_name(p: &Pat) -> Option<String> {
-        if let Pat::Ident(b) = p {
-            Some(b.id.sym.to_string())
-        } else {
-            None
-        }
-    }
-
-    fn body_has_return_value(stmts: &[Stmt]) -> bool {
-        for s in stmts {
-            if has_return_value_stmt(s) {
-                return true;
-            }
-        }
-        false
-    }
-    fn has_return_value_stmt(s: &Stmt) -> bool {
-        match s {
-            Stmt::Return(r) => r.arg.is_some(),
-            Stmt::Block(b) => body_has_return_value(&b.stmts),
-            Stmt::If(i) => {
-                has_return_value_stmt(&i.cons)
-                    || i.alt.as_deref().map_or(false, has_return_value_stmt)
-            }
-            Stmt::While(w) => has_return_value_stmt(&w.body),
-            Stmt::DoWhile(w) => has_return_value_stmt(&w.body),
-            Stmt::For(f) => has_return_value_stmt(&f.body),
-            Stmt::ForOf(f) => has_return_value_stmt(&f.body),
-            Stmt::ForIn(f) => has_return_value_stmt(&f.body),
-            Stmt::Try(t) => {
-                body_has_return_value(&t.block.stmts)
-                    || t.handler
-                        .as_ref()
-                        .map_or(false, |h| body_has_return_value(&h.body.stmts))
-                    || t.finalizer
-                        .as_ref()
-                        .map_or(false, |f| body_has_return_value(&f.stmts))
-            }
-            _ => false,
-        }
-    }
-
-    fn build_fn_decl(
-        name: &str,
-        params: &[swc_ecma_ast::Param],
-        body_stmts: Vec<Stmt>,
-    ) -> FunctionDecl {
-        // (#object_builtins) Suporte a destructuring em params de arrow/fn
-        // hoisted: \`forEach(([k,v]) => ...)\`. Geramos param sintetico por
-        // posicao e prepend \`const [k,v] = __destruct_param_N;\` ao body.
-        let mut prologue: Vec<Stmt> = Vec::new();
-        let mut parameters: Vec<Parameter> = Vec::with_capacity(params.len());
-        for (i, p) in params.iter().enumerate() {
-            let pat_for_destruct: Option<&Pat> = match &p.pat {
-                Pat::Array(_) | Pat::Object(_) => Some(&p.pat),
-                Pat::Assign(a) if matches!(a.left.as_ref(), Pat::Array(_) | Pat::Object(_)) => {
-                    Some(a.left.as_ref())
-                }
-                _ => None,
-            };
-            if let Some(pat) = pat_for_destruct {
-                let synth_name = format!("__hoist_destruct_{}_{}", name, i);
-                // const <pat> = <synth>;
-                let synth_ident = swc_ecma_ast::Ident::new(
-                    synth_name.as_str().into(),
-                    Default::default(),
-                    Default::default(),
-                );
-                let init = Expr::Ident(synth_ident);
-                let var_decl = swc_ecma_ast::VarDecl {
-                    span: Default::default(),
-                    ctxt: Default::default(),
-                    kind: swc_ecma_ast::VarDeclKind::Const,
-                    declare: false,
-                    decls: vec![swc_ecma_ast::VarDeclarator {
-                        span: Default::default(),
-                        name: pat.clone(),
-                        init: Some(Box::new(init)),
-                        definite: false,
-                    }],
-                };
-                prologue.push(Stmt::Decl(swc_ecma_ast::Decl::Var(Box::new(var_decl))));
-                parameters.push(Parameter {
-                    name: synth_name,
-                    type_annotation: None,
-                    modifiers: MemberModifiers::default(),
-                    variadic: false,
-                    default: None,
-                    span: Span::default(),
-                });
-            } else if let Some(n) = pat_to_param_name(&p.pat) {
-                parameters.push(Parameter {
-                    name: n,
-                    type_annotation: None,
-                    modifiers: MemberModifiers::default(),
-                    variadic: false,
-                    default: None,
-                    span: Span::default(),
-                });
-            }
-        }
-        // Heuristica: se body tem \`return <expr>\`, declaramos retorno
-        // i64. Sem isso, declare_user_fn vira void e o codegen descarta
-        // o valor retornado, quebrando \`apply(fn, x)\` e IIFE com retorno.
-        let return_type = if body_has_return_value(&body_stmts) {
-            Some("i64".to_string())
-        } else {
-            None
-        };
-        let mut all_stmts: Vec<Stmt> = prologue;
-        all_stmts.extend(body_stmts);
-        let body: Vec<Statement> = all_stmts
-            .into_iter()
-            .map(|s| {
-                Statement::Raw(
-                    RawStmt::new("<hoisted-fn>".to_string(), Span::default()).with_stmt(s),
-                )
-            })
-            .collect();
-        FunctionDecl {
-            name: name.to_string(),
-            parameters,
-            return_type,
-            body,
-            span: Span::default(),
-            is_async: false,
-        }
-    }
-
-    fn arrow_body_stmts(arrow: &ArrowExpr) -> Vec<Stmt> {
-        match arrow.body.as_ref() {
-            BlockStmtOrExpr::BlockStmt(b) => b.stmts.clone(),
-            BlockStmtOrExpr::Expr(e) => vec![Stmt::Return(swc_ecma_ast::ReturnStmt {
-                span: arrow.span,
-                arg: Some(e.clone()),
-            })],
-        }
-    }
-
-    fn arrow_params(arrow: &ArrowExpr) -> Vec<swc_ecma_ast::Param> {
-        arrow
-            .params
-            .iter()
-            .map(|pat| swc_ecma_ast::Param {
-                span: Default::default(),
-                decorators: Vec::new(),
-                pat: pat.clone(),
-            })
-            .collect()
-    }
-
-    fn visit_expr(expr: &mut Expr, counter: &mut u32, new_fns: &mut Vec<Item>) {
-        // Pre-order: primeiro visitamos sub-exprs (para que Fn/Arrow
-        // aninhadas tambem sejam hoisted) e depois substituimos a
-        // raiz se for Fn/Arrow.
-        descend_expr(expr, counter, new_fns);
-
-        // Peel \`(expr)\` wrappers — IIFE \`(function(){...})(...)\` apos
-        // hoist vira \`(__hoisted_fn_N)(...)\` e o codegen so' reconhece
-        // Expr::Ident como callee direto.
-        loop {
-            let take = matches!(expr, Expr::Paren(_));
-            if !take {
-                break;
-            }
-            if let Expr::Paren(p) = std::mem::replace(
-                expr,
-                Expr::Invalid(swc_ecma_ast::Invalid {
-                    span: Default::default(),
-                }),
-            ) {
-                *expr = *p.expr;
-            }
-        }
-
-        let replace = match expr {
-            Expr::Fn(_) | Expr::Arrow(_) => true,
-            _ => false,
-        };
-        if !replace {
-            return;
-        }
-
-        let owned = std::mem::replace(
-            expr,
-            Expr::Invalid(swc_ecma_ast::Invalid {
-                span: Default::default(),
-            }),
-        );
-        let is_arrow_expr = matches!(owned, Expr::Arrow(_));
-        let name = if is_arrow_expr {
-            format!("__hoisted_arrow_{}", *counter)
-        } else {
-            format!("__hoisted_fn_{}", *counter)
-        };
-        *counter += 1;
-
-        let fn_decl = match owned {
-            Expr::Fn(fn_expr) => {
-                let func = &fn_expr.function;
-                let body = func.body.as_ref().map(|b| b.stmts.clone()).unwrap_or_default();
-                build_fn_decl(&name, &func.params, body)
-            }
-            Expr::Arrow(arrow) => {
-                let params = arrow_params(&arrow);
-                let body = arrow_body_stmts(&arrow);
-                build_fn_decl(&name, &params, body)
-            }
-            _ => unreachable!(),
-        };
-        new_fns.push(Item::Function(fn_decl));
-        *expr = fresh_ident(&name);
-    }
-
-    fn descend_expr(expr: &mut Expr, counter: &mut u32, new_fns: &mut Vec<Item>) {
-        match expr {
-            Expr::Call(c) => {
-                if let Callee::Expr(callee) = &mut c.callee {
-                    visit_expr(callee.as_mut(), counter, new_fns);
-                }
-                for arg in c.args.iter_mut() {
-                    visit_expr(arg.expr.as_mut(), counter, new_fns);
-                }
-            }
-            Expr::New(n) => {
-                visit_expr(n.callee.as_mut(), counter, new_fns);
-                if let Some(args) = n.args.as_mut() {
-                    for arg in args.iter_mut() {
-                        visit_expr(arg.expr.as_mut(), counter, new_fns);
-                    }
-                }
-            }
-            Expr::Bin(b) => {
-                visit_expr(b.left.as_mut(), counter, new_fns);
-                visit_expr(b.right.as_mut(), counter, new_fns);
-            }
-            Expr::Assign(a) => {
-                visit_expr(a.right.as_mut(), counter, new_fns);
-            }
-            Expr::Unary(u) => visit_expr(u.arg.as_mut(), counter, new_fns),
-            Expr::Update(u) => visit_expr(u.arg.as_mut(), counter, new_fns),
-            Expr::Cond(c) => {
-                visit_expr(c.test.as_mut(), counter, new_fns);
-                visit_expr(c.cons.as_mut(), counter, new_fns);
-                visit_expr(c.alt.as_mut(), counter, new_fns);
-            }
-            Expr::Seq(s) => {
-                for e in s.exprs.iter_mut() {
-                    visit_expr(e.as_mut(), counter, new_fns);
-                }
-            }
-            Expr::Member(m) => {
-                visit_expr(m.obj.as_mut(), counter, new_fns);
-                if let MemberProp::Computed(c) = &mut m.prop {
-                    visit_expr(c.expr.as_mut(), counter, new_fns);
-                }
-            }
-            Expr::OptChain(o) => {
-                use swc_ecma_ast::OptChainBase;
-                match o.base.as_mut() {
-                    OptChainBase::Member(m) => {
-                        visit_expr(m.obj.as_mut(), counter, new_fns);
-                    }
-                    OptChainBase::Call(c) => {
-                        visit_expr(c.callee.as_mut(), counter, new_fns);
-                        for arg in c.args.iter_mut() {
-                            visit_expr(arg.expr.as_mut(), counter, new_fns);
-                        }
-                    }
-                }
-            }
-            Expr::Array(arr) => {
-                for elem in arr.elems.iter_mut().flatten() {
-                    visit_expr(elem.expr.as_mut(), counter, new_fns);
-                }
-            }
-            Expr::Object(obj) => {
-                for prop in obj.props.iter_mut() {
-                    if let swc_ecma_ast::PropOrSpread::Prop(p) = prop {
-                        if let swc_ecma_ast::Prop::KeyValue(kv) = p.as_mut() {
-                            visit_expr(kv.value.as_mut(), counter, new_fns);
-                        }
-                    }
-                }
-            }
-            Expr::Tpl(t) => {
-                for e in t.exprs.iter_mut() {
-                    visit_expr(e.as_mut(), counter, new_fns);
-                }
-            }
-            Expr::Paren(p) => visit_expr(p.expr.as_mut(), counter, new_fns),
-            Expr::TsAs(a) => visit_expr(a.expr.as_mut(), counter, new_fns),
-            Expr::TsTypeAssertion(a) => visit_expr(a.expr.as_mut(), counter, new_fns),
-            Expr::TsConstAssertion(a) => visit_expr(a.expr.as_mut(), counter, new_fns),
-            Expr::TsSatisfies(a) => visit_expr(a.expr.as_mut(), counter, new_fns),
-            Expr::TsNonNull(n) => visit_expr(n.expr.as_mut(), counter, new_fns),
-            Expr::Await(a) => visit_expr(a.arg.as_mut(), counter, new_fns),
-            Expr::Yield(y) => {
-                if let Some(e) = y.arg.as_mut() {
-                    visit_expr(e.as_mut(), counter, new_fns);
-                }
-            }
-            // Fn/Arrow: nao descer dentro do body — body sera tratado
-            // quando o pass rodar de novo recursivamente nos new_fns.
-            // Aqui apenas paramos. Se body tiver mais lambdas, elas
-            // sao hoisted no proximo loop fixed-point.
-            _ => {}
-        }
-    }
-
-    fn visit_stmt(stmt: &mut Stmt, counter: &mut u32, new_fns: &mut Vec<Item>) {
-        match stmt {
-            Stmt::Expr(e) => visit_expr(e.expr.as_mut(), counter, new_fns),
-            Stmt::Return(r) => {
-                if let Some(e) = r.arg.as_mut() {
-                    visit_expr(e.as_mut(), counter, new_fns);
-                }
-            }
-            Stmt::Decl(Decl::Var(vd)) => {
-                for d in vd.decls.iter_mut() {
-                    if let Some(init) = d.init.as_mut() {
-                        visit_expr(init.as_mut(), counter, new_fns);
-                    }
-                }
-            }
-            Stmt::Block(b) => {
-                for s in b.stmts.iter_mut() {
-                    visit_stmt(s, counter, new_fns);
-                }
-            }
-            Stmt::If(i) => {
-                visit_expr(i.test.as_mut(), counter, new_fns);
-                visit_stmt(i.cons.as_mut(), counter, new_fns);
-                if let Some(alt) = i.alt.as_mut() {
-                    visit_stmt(alt.as_mut(), counter, new_fns);
-                }
-            }
-            Stmt::While(w) => {
-                visit_expr(w.test.as_mut(), counter, new_fns);
-                visit_stmt(w.body.as_mut(), counter, new_fns);
-            }
-            Stmt::DoWhile(d) => {
-                visit_expr(d.test.as_mut(), counter, new_fns);
-                visit_stmt(d.body.as_mut(), counter, new_fns);
-            }
-            Stmt::For(f) => {
-                if let Some(init) = f.init.as_mut() {
-                    if let swc_ecma_ast::VarDeclOrExpr::Expr(e) = init {
-                        visit_expr(e.as_mut(), counter, new_fns);
-                    }
-                    if let swc_ecma_ast::VarDeclOrExpr::VarDecl(vd) = init {
-                        for d in vd.decls.iter_mut() {
-                            if let Some(init) = d.init.as_mut() {
-                                visit_expr(init.as_mut(), counter, new_fns);
-                            }
-                        }
-                    }
-                }
-                if let Some(t) = f.test.as_mut() {
-                    visit_expr(t.as_mut(), counter, new_fns);
-                }
-                if let Some(u) = f.update.as_mut() {
-                    visit_expr(u.as_mut(), counter, new_fns);
-                }
-                visit_stmt(f.body.as_mut(), counter, new_fns);
-            }
-            Stmt::ForOf(f) => {
-                visit_expr(f.right.as_mut(), counter, new_fns);
-                visit_stmt(f.body.as_mut(), counter, new_fns);
-            }
-            Stmt::ForIn(f) => {
-                visit_expr(f.right.as_mut(), counter, new_fns);
-                visit_stmt(f.body.as_mut(), counter, new_fns);
-            }
-            Stmt::Switch(s) => {
-                visit_expr(s.discriminant.as_mut(), counter, new_fns);
-                for c in s.cases.iter_mut() {
-                    if let Some(t) = c.test.as_mut() {
-                        visit_expr(t.as_mut(), counter, new_fns);
-                    }
-                    for s in c.cons.iter_mut() {
-                        visit_stmt(s, counter, new_fns);
-                    }
-                }
-            }
-            Stmt::Throw(t) => visit_expr(t.arg.as_mut(), counter, new_fns),
-            Stmt::Try(t) => {
-                for s in t.block.stmts.iter_mut() {
-                    visit_stmt(s, counter, new_fns);
-                }
-                if let Some(handler) = t.handler.as_mut() {
-                    for s in handler.body.stmts.iter_mut() {
-                        visit_stmt(s, counter, new_fns);
-                    }
-                }
-                if let Some(finalizer) = t.finalizer.as_mut() {
-                    for s in finalizer.stmts.iter_mut() {
-                        visit_stmt(s, counter, new_fns);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    // Iterate ate fixed-point: visitar items, hoistar, e depois
-    // re-visitar new_fns (que podem conter outras Fn/Arrow aninhadas).
-    let mut to_visit_items: Vec<&mut Item> = program.items.iter_mut().collect();
-    for item in to_visit_items.drain(..) {
-        match item {
-            Item::Function(f) => {
-                for s in f.body.iter_mut() {
-                    let Statement::Raw(raw) = s;
-                    if let Some(stmt) = raw.stmt.as_mut() {
-                        visit_stmt(stmt, &mut counter, &mut new_fns);
-                    }
-                }
-            }
-            Item::Statement(Statement::Raw(raw)) => {
-                if let Some(stmt) = raw.stmt.as_mut() {
-                    visit_stmt(stmt, &mut counter, &mut new_fns);
-                }
-            }
-            Item::Class(class) => {
-                for member in class.members.iter_mut() {
-                    let body = match member {
-                        ClassMember::Method(m) => &mut m.body,
-                        ClassMember::Constructor(c) => &mut c.body,
-                        _ => continue,
-                    };
-                    for s in body.iter_mut() {
-                        let Statement::Raw(raw) = s;
-                        if let Some(stmt) = raw.stmt.as_mut() {
-                            visit_stmt(stmt, &mut counter, &mut new_fns);
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    // Process new_fns recursivamente (lambdas dentro de lambdas).
-    let mut work = std::mem::take(&mut new_fns);
-    let mut acc: Vec<Item> = Vec::new();
-    while let Some(item) = work.pop() {
-        if let Item::Function(mut f) = item {
-            for s in f.body.iter_mut() {
-                let Statement::Raw(raw) = s;
-                if let Some(stmt) = raw.stmt.as_mut() {
-                    let mut local_new: Vec<Item> = Vec::new();
-                    visit_stmt(stmt, &mut counter, &mut local_new);
-                    work.extend(local_new);
-                }
-            }
-            acc.push(Item::Function(f));
-        } else {
-            acc.push(item);
-        }
-    }
-    program.items.extend(acc);
-}
-
-fn expand_destructuring(program: &mut Program) {
-    // Counter global para gerar nomes de temporários únicos.
-    let mut counter: u32 = 0;
-
-    // Helper: processa um Vec<Statement>, expandindo decls de destructuring.
-    // Cria novos statements e replace in place.
-    fn process_body(body: &mut Vec<Statement>, counter: &mut u32) {
-        let mut i = 0;
-        while i < body.len() {
-            let Statement::Raw(raw) = &body[i];
-            let Some(stmt) = raw.stmt.as_ref() else {
-                i += 1;
-                continue;
-            };
-            // Apenas \`Stmt::Decl(Decl::Var)\` pode ter patterns.
-            let Stmt::Decl(Decl::Var(var_decl)) = stmt else {
-                i += 1;
-                continue;
-            };
-
-            // Verifica se algum decl tem pattern.
-            let has_pattern = var_decl
-                .decls
-                .iter()
-                .any(|d| !matches!(d.name, Pat::Ident(_)));
-            if !has_pattern {
-                i += 1;
-                continue;
-            }
-
-            // Expande: gera múltiplos statements.
-            let kind = var_decl.kind;
-            let mut new_stmts: Vec<Statement> = Vec::new();
-            for decl in &var_decl.decls {
-                expand_destruct_decl(
-                    &decl.name,
-                    decl.init.as_deref(),
-                    kind,
-                    counter,
-                    &mut new_stmts,
-                );
-            }
-
-            // Substitui o stmt atual pelos novos.
-            body.remove(i);
-            for (k, s) in new_stmts.into_iter().enumerate() {
-                body.insert(i + k, s);
-            }
-            // Skip pelos stmts inseridos (eles já são "Ident" decls,
-            // não recursam aqui).
-            // i fica no próximo após o último inserido — mas como
-            // recursamos via process_body em fns aninhadas só, ok.
-            // Avança pelo número de inserts.
-            // (Já incrementei body.remove + insert; i agora aponta pro
-            // primeiro inserido. Avanço pra depois deles.)
-            // Não faço skip extra pq são todos Ident decls.
-            i += 1;
-        }
-    }
-
-    for item in program.items.iter_mut() {
-        match item {
-            Item::Function(f) => process_body(&mut f.body, &mut counter),
-            Item::Class(c) => {
-                for m in c.members.iter_mut() {
-                    match m {
-                        ClassMember::Constructor(ctor) => {
-                            process_body(&mut ctor.body, &mut counter);
-                        }
-                        ClassMember::Method(method) => {
-                            process_body(&mut method.body, &mut counter);
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            Item::Statement(_) => {
-                // Top-level: empacota num Vec único e processa.
-                // A reorganização é mais delicada porque os items são
-                // Item::Statement, não Statement. Simplifico: detecta
-                // patterns em items individuais.
-            }
-            _ => {}
-        }
-    }
-
-    // Top-level: itera com índice mutável.
-    let mut i = 0;
-    while i < program.items.len() {
-        let needs_expansion = matches!(
-            &program.items[i],
-            Item::Statement(Statement::Raw(raw))
-                if matches!(
-                    raw.stmt.as_ref(),
-                    Some(Stmt::Decl(Decl::Var(v)))
-                    if v.decls.iter().any(|d| !matches!(d.name, Pat::Ident(_)))
-                )
-        );
-        if !needs_expansion {
-            i += 1;
-            continue;
-        }
-        let Item::Statement(Statement::Raw(raw)) = &program.items[i] else {
-            i += 1;
-            continue;
-        };
-        let Some(Stmt::Decl(Decl::Var(var_decl))) = raw.stmt.as_ref() else {
-            i += 1;
-            continue;
-        };
-        let kind = var_decl.kind;
-        let mut new_stmts: Vec<Statement> = Vec::new();
-        for decl in &var_decl.decls {
-            expand_destruct_decl(
-                &decl.name,
-                decl.init.as_deref(),
-                kind,
-                &mut counter,
-                &mut new_stmts,
-            );
-        }
-        program.items.remove(i);
-        for (k, s) in new_stmts.into_iter().enumerate() {
-            program.items.insert(i + k, Item::Statement(s));
-        }
-        // Avança pelos inseridos.
-        // (não faço i += new_stmts.len() porque já consumi via into_iter,
-        // mas acima o número era inserido sequencialmente; após insert,
-        // i deve avançar pelo número de elementos inseridos. Loop simples:
-        // body[i..i+N] são novos Ident decls, não precisam re-expansão.)
-        // Conta quantos foram inseridos.
-        // Como não temos esse número aqui (consumiu o iter), recomputo
-        // varredando até achar não-Statement ou seja arbitrário.
-        // Mais simples: continuar do índice atual; os novos decls são
-        // Ident (não disparam needs_expansion).
-        // Sem `i +=`, pq remove + insert deixa i no primeiro inserido.
-        i += 1;
-    }
-}
-
-/// Expande um decl com pattern para Vec<Statement>. Recurse em nested.
-fn expand_destruct_decl(
-    pat: &Pat,
-    init: Option<&Expr>,
-    kind: swc_ecma_ast::VarDeclKind,
-    counter: &mut u32,
-    out: &mut Vec<Statement>,
-) {
-    match pat {
-        Pat::Ident(_) => {
-            // Sem destructuring — apenas regenera o decl simples.
-            let var_decl = swc_ecma_ast::VarDecl {
-                span: Default::default(),
-                ctxt: Default::default(),
-                kind,
-                declare: false,
-                decls: vec![swc_ecma_ast::VarDeclarator {
-                    span: Default::default(),
-                    name: pat.clone(),
-                    init: init.map(|e| Box::new(e.clone())),
-                    definite: false,
-                }],
-            };
-            let stmt = Stmt::Decl(Decl::Var(Box::new(var_decl)));
-            out.push(Statement::Raw(
-                RawStmt::new("<destruct>".to_string(), Span::default()).with_stmt(stmt),
-            ));
-        }
-        Pat::Array(arr) => {
-            let tmp_name = format!("__destruct_{}", *counter);
-            *counter += 1;
-            // Gera const __destruct_N = init;
-            if let Some(init) = init {
-                out.push(make_const_decl(&tmp_name, init.clone(), kind));
-            }
-            for (idx, elem) in arr.elems.iter().enumerate() {
-                let Some(e) = elem else { continue }; // hole — pula
-
-                // Rest element: copia tudo a partir de idx para um novo
-                // array. Sem default ou nesting interno (caso atipico).
-                if let Pat::Rest(rest_pat) = e {
-                    let rest_name = match rest_pat.arg.as_ref() {
-                        Pat::Ident(b) => b.id.sym.to_string(),
-                        _ => continue,
-                    };
-                    push_array_rest_decl(&tmp_name, &rest_name, idx, kind, out);
-                    break; // rest precisa ser ultimo
-                }
-
-                // Default em array: \`[a = 5]\` — Pat::Assign.
-                if let Pat::Assign(ap) = e {
-                    let access = make_index_access_with_default(
-                        &tmp_name,
-                        idx as f64,
-                        ap.right.as_ref(),
-                    );
-                    expand_destruct_decl(ap.left.as_ref(), Some(&access), kind, counter, out);
-                    continue;
-                }
-
-                // const elem_name = __destruct_N[idx];
-                let access = make_index_access(&tmp_name, idx as f64);
-                expand_destruct_decl(&e, Some(&access), kind, counter, out);
-            }
-        }
-        Pat::Object(obj) => {
-            let tmp_name = format!("__destruct_{}", *counter);
-            *counter += 1;
-            if let Some(init) = init {
-                out.push(make_const_decl(&tmp_name, init.clone(), kind));
-            }
-            // Coleta keys explicitamente capturadas — usadas pra construir
-            // o `rest` removendo-as do clone (#312).
-            let mut consumed_keys: Vec<String> = Vec::new();
-            let mut rest_pat: Option<&Pat> = None;
-            for prop in &obj.props {
-                match prop {
-                    swc_ecma_ast::ObjectPatProp::Assign(ap) => {
-                        // \`{ x }\` ou \`{ x = default }\`
-                        let key = ap.key.id.sym.to_string();
-                        consumed_keys.push(key.clone());
-                        let inner_pat = Pat::Ident(swc_ecma_ast::BindingIdent {
-                            id: ap.key.id.clone(),
-                            type_ann: None,
-                        });
-                        let access = if let Some(default) = ap.value.as_ref() {
-                            make_member_access_with_default(&tmp_name, &key, default.as_ref())
-                        } else {
-                            make_member_access(&tmp_name, &key)
-                        };
-                        expand_destruct_decl(&inner_pat, Some(&access), kind, counter, out);
-                    }
-                    swc_ecma_ast::ObjectPatProp::KeyValue(kvp) => {
-                        // \`{ x: a }\` — alias; tambem \`{ x: a = default }\`
-                        // (#551) SWC representa default em KeyValue como
-                        // value=Pat::Assign{left,right=default}.
-                        let key = match &kvp.key {
-                            swc_ecma_ast::PropName::Ident(id) => id.sym.to_string(),
-                            swc_ecma_ast::PropName::Str(s) => s.value.to_string_lossy().to_string(),
-                            _ => continue,
-                        };
-                        consumed_keys.push(key.clone());
-                        let (target_pat, access) = if let Pat::Assign(ap) = kvp.value.as_ref() {
-                            let acc = make_member_access_with_default(
-                                &tmp_name,
-                                &key,
-                                ap.right.as_ref(),
-                            );
-                            (ap.left.as_ref(), acc)
-                        } else {
-                            let acc = make_member_access(&tmp_name, &key);
-                            (kvp.value.as_ref(), acc)
-                        };
-                        expand_destruct_decl(target_pat, Some(&access), kind, counter, out);
-                    }
-                    swc_ecma_ast::ObjectPatProp::Rest(rest) => {
-                        rest_pat = Some(rest.arg.as_ref());
-                    }
-                }
-            }
-            // Object rest (#312): `const { a, b, ...rest } = obj` →
-            //   const rest = collections.map_clone(__destruct_N);
-            //   collections.map_delete(rest, "a");
-            //   collections.map_delete(rest, "b");
-            if let Some(pat) = rest_pat {
-                if let Pat::Ident(rest_id) = pat {
-                    let rest_name = rest_id.id.sym.to_string();
-                    push_object_rest_decls(&rest_name, &tmp_name, &consumed_keys, kind, out);
-                }
-            }
-        }
-        _ => {
-            // Outros patterns (Rest, Assign solo, etc) — emite decl direto
-            // se for Ident interno; senão silencia.
-        }
-    }
-}
-
-/// `const <name> = <expr>;` (kind preservado).
-fn make_const_decl(name: &str, expr: Expr, kind: swc_ecma_ast::VarDeclKind) -> Statement {
-    let var = swc_ecma_ast::VarDecl {
-        span: Default::default(),
-        ctxt: Default::default(),
-        kind,
-        declare: false,
-        decls: vec![swc_ecma_ast::VarDeclarator {
-            span: Default::default(),
-            name: Pat::Ident(swc_ecma_ast::BindingIdent {
-                id: swc_ecma_ast::Ident {
-                    span: Default::default(),
-                    ctxt: Default::default(),
-                    sym: name.into(),
-                    optional: false,
-                },
-                type_ann: None,
-            }),
-            init: Some(Box::new(expr)),
-            definite: false,
-        }],
-    };
-    let stmt = Stmt::Decl(Decl::Var(Box::new(var)));
-    Statement::Raw(RawStmt::new("<destruct-tmp>".to_string(), Span::default()).with_stmt(stmt))
-}
-
-/// `<obj>[<idx>]` como Expr.
-fn make_index_access(obj_name: &str, idx: f64) -> Expr {
-    Expr::Member(swc_ecma_ast::MemberExpr {
-        span: Default::default(),
-        obj: Box::new(Expr::Ident(swc_ecma_ast::Ident {
-            span: Default::default(),
-            ctxt: Default::default(),
-            sym: obj_name.into(),
-            optional: false,
-        })),
-        prop: MemberProp::Computed(swc_ecma_ast::ComputedPropName {
-            span: Default::default(),
-            expr: Box::new(Expr::Lit(Lit::Num(swc_ecma_ast::Number {
-                span: Default::default(),
-                value: idx,
-                raw: Some(format!("{}", idx as i64).into()),
-            }))),
-        }),
-    })
-}
-
-/// `<obj>.<key>` como Expr.
-fn make_member_access(obj_name: &str, key: &str) -> Expr {
-    Expr::Member(swc_ecma_ast::MemberExpr {
-        span: Default::default(),
-        obj: Box::new(Expr::Ident(swc_ecma_ast::Ident {
-            span: Default::default(),
-            ctxt: Default::default(),
-            sym: obj_name.into(),
-            optional: false,
-        })),
-        prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-            span: Default::default(),
-            sym: key.into(),
-        }),
-    })
-}
-
-/// `<obj>[<idx>] ?? <default>` como Expr.
-/// Em destructuring com default, queremos: se o slot eh nullish/missing,
-/// usa default. Em RTS isso tambem dispara em valor 0 — JS distingue mas
-/// custaria adicionar IR proprio so pra isso. Boa aproximacao pra MVP.
-fn make_index_access_with_default(obj_name: &str, idx: f64, default: &Expr) -> Expr {
-    let access = make_index_access(obj_name, idx);
-    Expr::Bin(swc_ecma_ast::BinExpr {
-        span: Default::default(),
-        op: swc_ecma_ast::BinaryOp::NullishCoalescing,
-        left: Box::new(access),
-        right: Box::new(default.clone()),
-    })
-}
-
-/// `<obj>.<key> ?? <default>` como Expr.
-fn make_member_access_with_default(obj_name: &str, key: &str, default: &Expr) -> Expr {
-    let access = make_member_access(obj_name, key);
-    Expr::Bin(swc_ecma_ast::BinExpr {
-        span: Default::default(),
-        op: swc_ecma_ast::BinaryOp::NullishCoalescing,
-        left: Box::new(access),
-        right: Box::new(default.clone()),
-    })
-}
-
-/// Empurra em `out` o desugar de object rest (#312):
-///   `const <rest_name> = collections.map_clone(<src_name>);`
-///   `collections.map_delete(<rest_name>, "<key>");` (uma por consumed_key)
-///
-/// Reusa o builtin `collections.map_clone` (registra entry novo no
-/// HandleTable com clone do IndexMap) + `map_delete` ja existente.
-fn push_object_rest_decls(
-    rest_name: &str,
-    src_name: &str,
-    consumed_keys: &[String],
-    kind: swc_ecma_ast::VarDeclKind,
-    out: &mut Vec<Statement>,
-) {
-    use swc_ecma_ast::{CallExpr, ExprOrSpread, ExprStmt, Ident, MemberExpr, Stmt};
-
-    fn ident(name: &str) -> Ident {
-        Ident {
-            span: Default::default(),
-            ctxt: Default::default(),
-            sym: name.into(),
-            optional: false,
-        }
-    }
-
-    // collections.map_clone(<src_name>)
-    let clone_call = Expr::Call(CallExpr {
-        span: Default::default(),
-        ctxt: Default::default(),
-        callee: swc_ecma_ast::Callee::Expr(Box::new(Expr::Member(MemberExpr {
-            span: Default::default(),
-            obj: Box::new(Expr::Ident(ident("collections"))),
-            prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-                span: Default::default(),
-                sym: "map_clone".into(),
-            }),
-        }))),
-        args: vec![ExprOrSpread {
-            spread: None,
-            expr: Box::new(Expr::Ident(ident(src_name))),
-        }],
-        type_args: None,
-    });
-    out.push(make_const_decl(rest_name, clone_call, kind));
-
-    // collections.map_delete(<rest_name>, "<key>"); para cada consumed_key
-    for key in consumed_keys {
-        let delete_call = Expr::Call(CallExpr {
-            span: Default::default(),
-            ctxt: Default::default(),
-            callee: swc_ecma_ast::Callee::Expr(Box::new(Expr::Member(MemberExpr {
-                span: Default::default(),
-                obj: Box::new(Expr::Ident(ident("collections"))),
-                prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-                    span: Default::default(),
-                    sym: "map_delete".into(),
-                }),
-            }))),
-            args: vec![
-                ExprOrSpread {
-                    spread: None,
-                    expr: Box::new(Expr::Ident(ident(rest_name))),
-                },
-                ExprOrSpread {
-                    spread: None,
-                    expr: Box::new(Expr::Lit(Lit::Str(swc_ecma_ast::Str {
-                        span: Default::default(),
-                        value: key.as_str().into(),
-                        raw: None,
-                    }))),
-                },
-            ],
-            type_args: None,
-        });
-        let stmt = Stmt::Expr(ExprStmt {
-            span: Default::default(),
-            expr: Box::new(delete_call),
-        });
-        let raw = crate::parser::ast::RawStmt::new(
-            format!("collections.map_delete({rest_name}, \"{key}\");"),
-            crate::parser::span::Span::default(),
-        )
-        .with_stmt(stmt);
-        out.push(crate::parser::ast::Statement::Raw(raw));
-    }
-}
-
-/// Empurra em \`out\`:
-///   \`const <rest_name> = [];\`
-///   \`for (let __i = <start_idx>; __i < <obj>.length; __i++) <rest_name>.push(<obj>[__i]);\`
-fn push_array_rest_decl(
-    obj_name: &str,
-    rest_name: &str,
-    start_idx: usize,
-    kind: swc_ecma_ast::VarDeclKind,
-    out: &mut Vec<Statement>,
-) {
-    use swc_ecma_ast::{
-        AssignOp, BinaryOp, BlockStmt, ForStmt, Ident, MemberExpr, Stmt, UpdateExpr, UpdateOp,
-        VarDecl, VarDeclKind, VarDeclOrExpr, VarDeclarator,
-    };
-
-    let counter_name = format!("__rest_i_{}", obj_name);
-
-    fn ident(name: &str) -> Ident {
-        Ident {
-            span: Default::default(),
-            ctxt: Default::default(),
-            sym: name.into(),
-            optional: false,
-        }
-    }
-
-    let arr_init = Expr::Array(swc_ecma_ast::ArrayLit {
-        span: Default::default(),
-        elems: vec![],
-    });
-
-    // const <rest_name> = [];
-    let rest_decl_stmt = make_const_decl(rest_name, arr_init, kind);
-
-    // for (let __rest_i = start_idx; __rest_i < obj.length; __rest_i++)
-    //   <rest_name>.push(obj[__rest_i]);
-    let init = VarDeclOrExpr::VarDecl(Box::new(VarDecl {
-        span: Default::default(),
-        ctxt: Default::default(),
-        kind: VarDeclKind::Let,
-        declare: false,
-        decls: vec![VarDeclarator {
-            span: Default::default(),
-            name: Pat::Ident(swc_ecma_ast::BindingIdent {
-                id: ident(&counter_name),
-                type_ann: None,
-            }),
-            init: Some(Box::new(Expr::Lit(Lit::Num(swc_ecma_ast::Number {
-                span: Default::default(),
-                value: start_idx as f64,
-                raw: Some(format!("{}", start_idx).into()),
-            })))),
-            definite: false,
-        }],
-    }));
-
-    let length_access = Expr::Member(MemberExpr {
-        span: Default::default(),
-        obj: Box::new(Expr::Ident(ident(obj_name))),
-        prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-            span: Default::default(),
-            sym: "length".into(),
-        }),
-    });
-
-    let test = Expr::Bin(swc_ecma_ast::BinExpr {
-        span: Default::default(),
-        op: BinaryOp::Lt,
-        left: Box::new(Expr::Ident(ident(&counter_name))),
-        right: Box::new(length_access),
-    });
-
-    let update = Expr::Update(UpdateExpr {
-        span: Default::default(),
-        op: UpdateOp::PlusPlus,
-        prefix: false,
-        arg: Box::new(Expr::Ident(ident(&counter_name))),
-    });
-
-    // <rest_name>.push(<obj>[<counter>])
-    let push_call = Expr::Call(swc_ecma_ast::CallExpr {
-        span: Default::default(),
-        ctxt: Default::default(),
-        callee: swc_ecma_ast::Callee::Expr(Box::new(Expr::Member(MemberExpr {
-            span: Default::default(),
-            obj: Box::new(Expr::Ident(ident(rest_name))),
-            prop: MemberProp::Ident(swc_ecma_ast::IdentName {
-                span: Default::default(),
-                sym: "push".into(),
-            }),
-        }))),
-        args: vec![swc_ecma_ast::ExprOrSpread {
-            spread: None,
-            expr: Box::new(Expr::Member(MemberExpr {
-                span: Default::default(),
-                obj: Box::new(Expr::Ident(ident(obj_name))),
-                prop: MemberProp::Computed(swc_ecma_ast::ComputedPropName {
-                    span: Default::default(),
-                    expr: Box::new(Expr::Ident(ident(&counter_name))),
-                }),
-            })),
-        }],
-        type_args: None,
-    });
-
-    let body_stmt = Stmt::Expr(swc_ecma_ast::ExprStmt {
-        span: Default::default(),
-        expr: Box::new(push_call),
-    });
-
-    let for_stmt = Stmt::For(ForStmt {
-        span: Default::default(),
-        init: Some(init),
-        test: Some(Box::new(test)),
-        update: Some(Box::new(update)),
-        body: Box::new(Stmt::Block(BlockStmt {
-            span: Default::default(),
-            ctxt: Default::default(),
-            stmts: vec![body_stmt],
-        })),
-    });
-
-    let _ = AssignOp::Assign;
-    out.push(rest_decl_stmt);
-    out.push(Statement::Raw(
-        RawStmt::new("<destruct-rest-loop>".to_string(), Span::default()).with_stmt(for_stmt),
-    ));
-}
 
 impl LiftAcc {
     /// Processa uma função user (não-classe, não-lifted). Detecta locais
