+++ crates/rts-codegen/src/codegen/lower/passes/mod.rs
@@ -7,5 +7,7 @@
 
 pub mod args;
 pub mod async_expand;
+pub mod destructuring;
+pub mod hoist_fn;
 pub mod object_methods;
 pub mod static_fields;
